requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
boto3==1.26.48
botocore==1.29.48
crispy-bootstrap5==0.7
Django==4.1.5
django-bootstrap-v5==1.0.11
django-crispy-forms==1.14.0
django-environ==0.9.0
django-storages==1.13.2
gunicorn==20.1.0
jmespath==1.0.1
mysqlclient==2.1.1
Pillow==9.4.0
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14